chore: work towards unified search
