+++ spog/ui/src/components/mod.rs
@@ -9,7 +9,6 @@ pub mod content;
 pub mod cvss;
 pub mod download;
 pub mod error;
-pub mod package;
 pub mod sbom;
 pub mod search;
 pub mod severity;
