+++ spog/ui/src/pages/sbom_search/mod.rs
@@ -1,38 +1,34 @@
 use std::rc::Rc;
 
-use crate::components::{common::PageHeading, package::PackageResult, sbom::CatalogSearch};
+use crate::components::{common::PageHeading, sbom::PackageResult, sbom::SbomSearch};
 use patternfly_yew::prelude::*;
 use spog_model::prelude::*;
 use yew::prelude::*;
 use yew_more_hooks::hooks::r#async::*;
 
 #[derive(Clone, Debug, PartialEq, Eq, Properties)]
-pub struct PackageProps {
+pub struct PackageProperties {
     #[prop_or_default]
     pub query: Option<String>,
 }
 
 #[function_component(Package)]
-pub fn package(props: &PackageProps) -> Html {
+pub fn package(props: &PackageProperties) -> Html {
     let search = use_state_eq(UseAsyncState::default);
-    let callback = {
-        let search = search.clone();
-        Callback::from(
-            move |state: UseAsyncHandleDeps<SearchResult<Rc<Vec<PackageSummary>>>, String>| {
-                search.set((*state).clone());
-            },
-        )
-    };
+    let callback = use_callback(
+        |state: UseAsyncHandleDeps<SearchResult<Rc<Vec<PackageSummary>>>, String>, search| search.set((*state).clone()),
+        search.clone(),
+    );
     let query = props.query.clone().filter(|s| !s.is_empty());
 
     html!(
         <>
-            <PageHeading subtitle="Search for SBOMs">{"Package Catalog"}</PageHeading>
+            <PageHeading subtitle="Search for SBOMs">{"SBOM Catalog"}</PageHeading>
 
             <PageSection variant={PageSectionVariant::Light}>
-                <CatalogSearch {callback} {query}>
+                <SbomSearch {callback} {query}>
                     <PackageResult state={(*search).clone()} />
-                </CatalogSearch>
+                </SbomSearch>
             </PageSection>
         </>
     )
